fix: rename `messages` package to `message` on commonJvmAndroid source set for `network` module
